flask_api/requirements.txt
Flask==0.12
Jinja2==2.9.5
MarkupSafe==0.23
Werkzeug==0.11.15
click==6.7
itsdangerous==0.24
requests==2.13.0
wsgiref==0.1.2
plaid-python
